networking adjustments
update info.plist for iOS, move trackoption changes from skalthoff, updates to login flows to allow overriding the protocol
