+++ src/providers/index.tsx
@@ -11,6 +11,7 @@ import { MMKVStorageKeys } from '../enums/mmkv-storage-keys'
 import { Api } from '@jellyfin/sdk/lib/api'
 import { JellyfinInfo } from '../api/info'
 import uuid from 'react-native-uuid'
+import { queryClient } from '../constants/query-client'
 
 /**
  * The context for the Jellify provider.
@@ -100,6 +101,10 @@ const JellifyContextInitializer = () => {
 		setServer(undefined)
 		setUser(undefined)
 		setLibrary(undefined)
+
+		queryClient.clear()
+
+		storage.clearAll()
 	}
 
 	useEffect(() => {
